+++ data_validation/query_builder/query_builder.py
@@ -96,6 +96,15 @@ class AggregateField(object):
             cast=cast,
         )
 
+    @staticmethod
+    def std(field_name=None, alias=None, cast=None):
+        return AggregateField(
+            ibis.expr.types.NumericColumn.std,
+            field_name=field_name,
+            alias=alias,
+            cast=cast,
+        )
+
     def compile(self, ibis_table):
         if self.field_name:
             agg_field = self.expr(ibis_table[self.field_name])
@@ -201,7 +210,7 @@ class ComparisonField(object):
         comparison_field = ibis_table[self.field_name]
         alias = self.alias or self.field_name
         if self.cast:
-            comparison_field = comparison_field.cast(self.cast)
+            comparison_field = comparison_field.force_cast(self.cast)
         comparison_field = comparison_field.name(alias)
 
         return comparison_field